Fix all -Wall warnings and enable -Werror
Mechanical (clean across Qt5/Qt6):
- 11× -Wformat=: %d → %lld + (long long) cast for qsizetype args
- 11× -Wsign-compare: cast loop bounds in xmlconv/xmlpreferences/
  mapcore/spawnshell/guildshell so uint indices and int/qsizetype
  sizes don't mix-sign-compare
- 6× -Wswitch on QVariant::Type (Qt5/Qt6 enum drift): cast switch
  operand to int so QMetaType case labels don't trip the enum check
- 5× -Wunused-but-set-variable: drop the variable, keep side effects
  (or gate on the same #ifdef as the only consumer)
- 1× -Wunused-function: drop legacy ok_in_hex helper

Real bugs:
- packet.cpp: setuid(getuid()) ignored its return value, so a setuid
  install would silently retain root if the syscall failed. Now exits.
- packetstream.cpp::connect2: payload was uninitialized when the
  EQPayloadList for an opcode had no entries; the subsequent
  `if (!payload)` read uninitialized memory. Default-init to nullptr.
- messageshell.cpp::spellFaded: read sf->message via offsetof so the
  flexible-array trailer stops triggering -Warray-bounds.

Build:
- Add -Werror to seq-daemon-core and showeq-daemon targets.
- Keep -Wno-unused-parameter / -Wno-deprecated-declarations (legacy
  surface, separate cleanup).
- -Wno-error=stringop-overread keeps Qt5's qstring.h false positive
  visible without breaking the Qt5 CI matrix.

Verified: ctest 3/3, replay 5/5 PASS, Qt5 + Qt6 builds clean.

+++ src/daemonapp.cpp
@@ -764,7 +764,7 @@ void DaemonApp::loadZoneMap(const QString& shortZoneName)
         }
         import = true;
     }
-    qInfo("loaded map for zone '%s' (%d layer(s) from %s)",
-          qUtf8Printable(shortZoneName), files.size(),
+    qInfo("loaded map for zone '%s' (%lld layer(s) from %s)",
+          qUtf8Printable(shortZoneName), (long long)files.size(),
           qUtf8Printable(QFileInfo(files.first()).absolutePath()));
 }
